refactor(control-api): co-locate policies tests into owning child modules
Move the 10 policy tests that touch child-private items out of the top-level
policies/mod.rs `tests` block into a `#[cfg(test)] mod tests` inside the module
that owns each tested item, so each item stays reachable from its test via the
module's own scope (no visibility widening):

- break-glass tests -> guard.rs (require_direct_policy_deploy_break_glass)
- impact/identity tests -> impact_validation/ (ensure_policy_proposal_deployable_impact,
  validate_policy_proposal_impact_matches_proposal, PolicyProposalRow)
- fleet rule-diff tests -> fleet_rule_diff/helpers.rs (reserve_policy_rule_diff_dispatch,
  policy_rule_diff_dispatch_response_action_ids,
  latest_policy_rule_diff_receipts_by_endpoint, CollectedPolicyRuleDiffReceipt)

The 3 policy_distribution contract tests (no policies-private access) stay in
mod.rs. Pure test relocation: all 13 tests still run and pass, assertions
unchanged.
