+++ crates/services/control-api/src/routes/policies/impact_validation/mod.rs
@@ -320,3 +320,105 @@ pub(crate) fn validate_policy_proposal_impact(
 
     Ok(impact)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn proposal_row_with_impact(impact: serde_json::Value) -> PolicyProposalRow {
+        let now = Utc::now();
+        PolicyProposalRow {
+            id: Uuid::from_u128(1),
+            tenant_id: Uuid::from_u128(2),
+            policy_yaml: "policy:\n  mode: test\n".to_string(),
+            checksum_sha256: "0".repeat(64),
+            description: None,
+            status: "pending".to_string(),
+            base_active_policy_version: 1,
+            proposed_policy_version: 2,
+            preview: serde_json::json!({}),
+            required_approvals: 1,
+            approved_by: Vec::new(),
+            approval_notes: serde_json::json!([]),
+            impact: Some(impact),
+            impact_attached_by: Some("test".to_string()),
+            impact_attached_at: Some(now),
+            deployed_policy_version: None,
+            deployment_id: None,
+            submitted_by: "test".to_string(),
+            reviewed_by: None,
+            review_note: None,
+            created_at: now,
+            updated_at: now,
+            reviewed_at: None,
+            deployed_at: None,
+        }
+    }
+
+    #[test]
+    fn deployable_impact_rejects_revise_recommendation() {
+        let proposal = proposal_row_with_impact(serde_json::json!({
+            "simulationReceiptsVerifiedCount": 1,
+            "blockingChangeCount": 0,
+            "recommendation": "revise"
+        }));
+
+        assert!(matches!(
+            ensure_policy_proposal_deployable_impact(&proposal),
+            Err(ApiError::Conflict(_))
+        ));
+    }
+
+    #[test]
+    fn deployable_impact_rejects_verified_blocking_changes() {
+        let proposal = proposal_row_with_impact(serde_json::json!({
+            "simulationReceiptsVerifiedCount": 1,
+            "blockingChangeCount": 1,
+            "recommendation": "observe_only"
+        }));
+
+        assert!(matches!(
+            ensure_policy_proposal_deployable_impact(&proposal),
+            Err(ApiError::Conflict(_))
+        ));
+    }
+
+    fn receipt_bound_impact(
+        proposed_policy_hash: &str,
+        proposed_policy_epoch: i64,
+    ) -> serde_json::Value {
+        serde_json::json!({
+            "simulationReceiptsVerifiedCount": 1,
+            "blockingChangeCount": 0,
+            "recommendation": "observe_only",
+            "simulationReceipts": [
+                {
+                    "evidenceValueHashes": {
+                        "proposedPolicyHash": sha256(proposed_policy_hash.as_bytes()).to_hex_prefixed(),
+                        "proposedPolicyEpoch": sha256(proposed_policy_epoch.to_string().as_bytes()).to_hex_prefixed()
+                    }
+                }
+            ]
+        })
+    }
+
+    #[test]
+    fn proposal_impact_receipts_must_match_proposal_policy_identity() {
+        let proposal = proposal_row_with_impact(receipt_bound_impact(&"0".repeat(64), 2));
+        let impact = receipt_bound_impact(&"1".repeat(64), 2);
+
+        assert!(matches!(
+            validate_policy_proposal_impact_matches_proposal(&proposal, &impact),
+            Err(ApiError::BadRequest(_))
+        ));
+    }
+
+    #[test]
+    fn proposal_impact_receipts_accept_matching_policy_identity() {
+        let proposal = proposal_row_with_impact(receipt_bound_impact(&"0".repeat(64), 2));
+        let impact = receipt_bound_impact(&"0".repeat(64), 2);
+
+        validate_policy_proposal_impact_matches_proposal(&proposal, &impact)
+            .expect("matching proposal receipt binding");
+    }
+}
